Backend/requirements.txt
flask
flask-cors
snowflake-connector-python>=3.0.0
pymupdf
python-pptx
requests
